requirements.txt
selenium==4.21.0
pillow==10.3.0